gymnax_flappy_2/pyproject.toml
[project]
name = "gymnax-flappy-2"
version = "0.1.0"
description = ""
authors = [
    {name = "Your Name",email = "you@example.com"}
]
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "jax (>=0.7.2,<0.8.0)",
    "flax (>=0.12.0,<0.13.0)",
    "pygame (>=2.6.1,<3.0.0)",
    "gymnax (>=0.0.9,<0.0.10)",
    "jax-rl @ git+https://github.com/hamishs/JAX-RL",
    "dm-haiku (>=0.0.15,<0.0.16)",
    "tensorflow-probability (>=0.25.0,<0.26.0)",
    "evosax (>=0.2.0,<0.3.0)",
    "tqdm (>=4.67.1,<5.0.0)"
]

[tool.poetry]
packages = [
    {include = "gymnax_flappy_2", from = "src"},
    {include = "gymnax_blines", from = "src"}
]

[build-system]
requires = ["poetry-core>=2.0.0,<3.0.0"]
build-backend = "poetry.core.masonry.api"
